Rename constructors.
