+++ crates/cli/src/commands/init.rs
@@ -247,14 +247,14 @@ async fn detect_vcs(dest_dir: &Path) -> Result<(String, String), AnyError> {
     if dest_dir.join(".git").exists() {
         return Ok((
             "git".into(),
-            Git::new("master", dest_dir)?.get_local_branch().await?,
+            Git::load("master", dest_dir)?.get_local_branch().await?,
         ));
     }
 
     if dest_dir.join(".svn").exists() {
         return Ok((
             "svn".into(),
-            Svn::new("trunk", dest_dir).get_local_branch().await?,
+            Svn::load("trunk", dest_dir).get_local_branch().await?,
         ));
     }
