+++ crates/emitter/benches/emitter_benchmark.rs
@@ -25,7 +25,7 @@ pub fn emit_benchmark(c: &mut Criterion) {
 
     c.bench_function("emitter_emit", |b| {
         b.to_async(&runtime).iter(|| async {
-            let workspace = Workspace::create(&workspace_root).await.unwrap();
+            let workspace = Workspace::load_from(&workspace_root).await.unwrap();
             let emitter = Emitter::new(Arc::new(RwLock::new(workspace)));
 
             emitter
