+++ crates/runner/benches/runner_benchmark.rs
@@ -9,7 +9,9 @@ use moon_workspace::Workspace;
 pub fn runner_benchmark(c: &mut Criterion) {
     c.bench_function("runner", |b| {
         b.iter(|| async {
-            let workspace = Workspace::create(get_fixtures_dir("cases")).await.unwrap();
+            let workspace = Workspace::load_from(get_fixtures_dir("cases"))
+                .await
+                .unwrap();
 
             let dep_graph = utils::setup_dep_graph(&workspace.projects);
 
@@ -26,7 +28,9 @@ pub fn runner_benchmark(c: &mut Criterion) {
 pub fn runner_with_platforms_benchmark(c: &mut Criterion) {
     c.bench_function("runner_with_platforms", |b| {
         b.iter(|| async {
-            let mut workspace = Workspace::create(get_fixtures_dir("cases")).await.unwrap();
+            let mut workspace = Workspace::load_from(get_fixtures_dir("cases"))
+                .await
+                .unwrap();
 
             utils::setup_platforms(&mut workspace);
