+++ crates/toolchain/tests/npm_test.rs
@@ -19,7 +19,7 @@ async fn create_npm_tool() -> (NodeTool, assert_fs::TempDir) {
         ..WorkspaceConfig::default()
     };
 
-    let toolchain = Toolchain::create_from(base_dir.path(), &config)
+    let toolchain = Toolchain::load_from(base_dir.path(), &config)
         .await
         .unwrap();
