+++ crates/toolchain/tests/toolchain_test.rs
@@ -12,7 +12,7 @@ async fn create_toolchain(base_dir: &Path) -> Toolchain {
         ..WorkspaceConfig::default()
     };
 
-    Toolchain::create_from(base_dir, &config).await.unwrap()
+    Toolchain::load_from(base_dir, &config).await.unwrap()
 }
 
 #[tokio::test]
