+++ crates/toolchain/tests/yarn_test.rs
@@ -21,7 +21,7 @@ async fn create_yarn_tool() -> (NodeTool, assert_fs::TempDir) {
         ..WorkspaceConfig::default()
     };
 
-    let toolchain = Toolchain::create_from(base_dir.path(), &config)
+    let toolchain = Toolchain::load_from(base_dir.path(), &config)
         .await
         .unwrap();
